FPS counter added to display
Messy, need a way to clean that up but it works, can remove the print console stuff.
Framerate is always arround 10% slower than the desired FPS, this is due to SDL delay being unable to delay any shorter than 1ms. not really an issue right now. Things to think about when working on multithreading.